REFACTOR: fixed all tests apart from 1.
