+++ frontend/src/test/Header.test.js
@@ -4,7 +4,7 @@ import { screen, act, render } from '@testing-library/react'
 import { BrowserRouter } from 'react-router-dom';
 
 
-import Header from './Header';
+import Header from '../Components/Header';
 
 describe('Header Snapshot tests',()=>{
     test(`Header matches snapshot`, () => {
